% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brunner_munzel_test.R
\name{brunner_munzel_test}
\alias{brunner_munzel_test}
\alias{brunner_munzel_test.default}
\alias{brunner_munzel_test.formula}
\title{Brunner-Munzel Test}
\source{
\url{https://github.com/vlyubchich/lawstat}
}
\usage{
brunner_munzel_test(...)

\method{brunner_munzel_test}{default}(x, y, alternative = c("two.sided",
  "greater", "less"), conf.level = 0.95, ...)

\method{brunner_munzel_test}{formula}(formula, data, ...)
}
\arguments{
\item{...}{further arguments to be passed to or from methods.}

\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{alternative}{a character string specifying the alternative
    hypothesis, must be one of \code{"two.sided"} (default),
    \code{"greater"} or \code{"less"}.  You can specify just the initial
    letter.}

\item{conf.level}{confidence level of the interval.}

\item{formula}{a formula of the form \code{lhs ~ rhs} where \code{lhs}
    is a numeric variable giving the data values and \code{rhs} a factor
    with two levels giving the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
    \code{\link{model.frame}}) containing the variables in the
    formula \code{formula}.  By default the variables are taken from
    \code{environment(formula)}.}
}
\description{
The original code was copied from lawstat v3.2 on 2019-01-23.
}
\seealso{
\code{stats::t.test}, \code{stats::wilcox.test}
}
