% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpoptime-package.R
\docType{package}
\name{cellpoptime-package}
\alias{cellpoptime}
\alias{cellpoptime-package}
\title{cellpoptime: analyze and visualize phylogenetic tree of cell population}
\description{
Analyze and visualize phylogenetic tree of cell population.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/cellpoptime}
  \item Report bugs at \url{https://github.com/heavywatal/cellpoptime/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavywatal@gmail.com} (\href{https://orcid.org/0000-0002-4718-7605}{ORCID})

}
\keyword{internal}
