% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpoptime.R
\docType{package}
\name{cellpoptime}
\alias{cellpoptime}
\alias{cellpoptime-package}
\title{cellpoptime: analyze and visualize phylogenetic tree of cell population}
\description{
cellpoptime: analyze and visualize phylogenetic tree of cell population
}
