% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{false_negative}
\alias{false_negative}
\alias{false_positive}
\title{Functions to judge false positive and false negative.}
\usage{
false_negative(predicted, actual)

false_positive(predicted, actual)
}
\arguments{
\item{predicted, actual}{vectors to compare}
}
\description{
Functions to judge false positive and false negative.
}
