% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{detect_driver}
\alias{detect_driver}
\alias{p_binom}
\title{Detect driver mutations}
\usage{
detect_driver(exp_desc)

p_binom(events)
}
\arguments{
\item{exp_desc}{expected number of descendants}

\item{events}{integer}
}
\description{
Detect driver mutations
}
