% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_root}
\alias{filter_root}
\alias{filter_children}
\alias{filter_offspring}
\alias{filter_clade}
\title{Filter tbl_tree by various criteria}
\usage{
filter_root(.data)

filter_children(.data, node)

filter_offspring(.data, node)

filter_clade(.data, node)
}
\arguments{
\item{.data}{tbl_tree}

\item{node}{integer ID}
}
\description{
Filter tbl_tree by various criteria
}
\details{
\code{filter_offspring} and \code{filter_clade} are similar functions;
the former does not include the specified node itself, but the latter does.
}
