% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify_cellpop}
\alias{fortify_cellpop}
\alias{ggtree_fortify}
\title{Convert tbl_tree to plottable data.frame}
\usage{
fortify_cellpop(model, data)

ggtree_fortify(data)
}
\arguments{
\item{model}{output of \code{scale_branches()}}

\item{data}{tbl_tree}
}
\description{
Convert tbl_tree to plottable data.frame
}
\details{
\code{fortify_cellpop} prepares plottable data.frame.

\code{ggtree_fortify} prepares plottable data.frame.
}
