% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{as_int_matrix}
\alias{as_int_matrix}
\alias{add_outgroup}
\title{Convert 01 string vector to integer matrix}
\usage{
as_int_matrix(x)

add_outgroup(mtrx)
}
\arguments{
\item{x}{string vector like '101'}

\item{mtrx}{integer matrix}
}
\value{
integer matrix

integer matrix
}
\description{
Convert 01 string vector to integer matrix

Add wild type row with genotype 000 and name 0
}
