% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_test.R
\name{permutation_test}
\alias{permutation_test}
\alias{permutation_test.default}
\alias{permutation_test.formula}
\title{Permutation Test}
\usage{
permutation_test(...)

\method{permutation_test}{default}(x, y, n = 1000L,
  alternative = c("two.sided", "greater", "less"))

\method{permutation_test}{formula}(formula, data, ...)
}
\arguments{
\item{...}{further arguments to be passed to or from methods.}

\item{x}{a (non-empty) numeric vector of data values.}

\item{y}{an optional (non-empty) numeric vector of data values.}

\item{n}{performs bootstrapping if n > 0}

\item{alternative}{a character string specifying the alternative
    hypothesis, must be one of \code{"two.sided"} (default),
    \code{"greater"} or \code{"less"}.  You can specify just the initial
    letter.}

\item{formula}{a formula of the form \code{lhs ~ rhs} where \code{lhs}
    is a numeric variable giving the data values and \code{rhs} a factor
    with two levels giving the corresponding groups.}

\item{data}{an optional matrix or data frame (or similar: see
    \code{\link{model.frame}}) containing the variables in the
    formula \code{formula}.  By default the variables are taken from
    \code{environment(formula)}.}
}
\description{
Permutation Test
}
\seealso{
\code{stats::t.test}, \code{stats::wilcox.test}
}
