% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_branches}
\alias{scale_branches}
\alias{scale_branches_record}
\title{Scale branches}
\usage{
scale_branches(x, detector = detect_driver_pois)

scale_branches_record(x, detector = detect_driver_pois)
}
\arguments{
\item{x}{data.frame}

\item{detector}{function}
}
\description{
Scale branches
}
\details{
\code{scale_branches} is a shortcut.

\code{scale_branches_record} records scaling process for explanation.
}
