% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellpoptime.R
\docType{package}
\name{cellpoptime-package}
\alias{cellpoptime}
\alias{cellpoptime-package}
\title{cellpoptime: analyze and visualize phylogenetic tree of cell population}
\description{
cellpoptime: analyze and visualize phylogenetic tree of cell population
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heavywatal/cellpoptime}
  \item Report bugs at \url{https://github.com/heavywatal/cellpoptime/issues}
}

}
\author{
\strong{Maintainer}: Watal M. Iwasaki \email{heavy.watal@gmail.com}

}
