% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{detect_driver}
\alias{detect_driver}
\title{Functions for detecting driver mutations
\code{detect_driver} returns a vector of p-values.}
\usage{
detect_driver(x)
}
\arguments{
\item{x}{a pair of branch lengths (number of mutations)}
}
\description{
Functions for detecting driver mutations
\code{detect_driver} returns a vector of p-values.
}
