% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{detect_driver}
\alias{detect_driver}
\alias{p_binom}
\title{Functions for detecting driver mutations
\code{detect_driver} returns a vector of p-values.}
\usage{
detect_driver(exp_desc)

p_binom(events)
}
\arguments{
\item{exp_desc}{expected number of descendants}

\item{events}{integer}
}
\description{
\code{p_binom} is a shortcut.
}
