% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{detect_driver}
\alias{detect_driver}
\alias{filter_origins}
\alias{mutant_labels}
\title{Functions for detecting driver mutations
\code{detect_driver} returns a vector of p-values.}
\usage{
detect_driver(v)

filter_origins(x, method = "fdr", q = 0.05)

mutant_labels(x)
}
\arguments{
\item{v}{a pair of branch lengths (number of mutations)}

\item{x}{scaled tbl_tree}

\item{method}{passed to \code{p.adjust()}}

\item{q}{threshold}
}
\description{
\code{mutant_labels} extracts mutated nodes in tbl_tree
}
