% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify.R
\name{fortify_cellpop}
\alias{fortify_cellpop}
\alias{ggtree_fortify}
\alias{fortify_rect}
\title{Convert tbl_tree to plottable data.frame}
\usage{
fortify_cellpop(model, data, method = "fdr", q = 0.05)

ggtree_fortify(data)

fortify_rect(data)
}
\arguments{
\item{model}{output of \code{scale_branches()}}

\item{data}{tbl_tree}

\item{method}{passed to \code{p.adjust()}}

\item{q}{threshold}
}
\description{
Convert tbl_tree to plottable data.frame
}
\details{
\code{fortify_cellpop} prepares plottable data.frame.

\code{ggtree_fortify} is a wrapper of \code{ggtree::fortify}.

\code{fortify_rect} calculates x-y coordinates in rectangular layout.
}
