% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{infer_tree}
\alias{infer_tree}
\title{Functions for genotype matrix}
\usage{
infer_tree(mtrx)
}
\arguments{
\item{mtrx}{integer matrix; rows are samples, columns are sites.}
}
\description{
\code{infer_tree} infers a phylogenetic tree from sampled genotypes.
}
