% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{infer_rooted_tree}
\alias{infer_rooted_tree}
\alias{add_outgroup}
\title{Functions for genotype matrix}
\usage{
infer_rooted_tree(mtrx)

add_outgroup(mtrx)
}
\arguments{
\item{mtrx}{integer matrix; rows are samples, columns are sites.}
}
\description{
\code{infer_rooted_tree} infers rooted tree of sampled genotypes
}
