% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fortify_cellpop}
\alias{fortify_cellpop}
\alias{plot_driver}
\title{Functions for visualization}
\usage{
fortify_cellpop(model, data, ...)

plot_driver(data, verbose = FALSE)
}
\arguments{
\item{model}{output of \code{scale_branches()}}

\item{data}{tbl_tree}

\item{...}{unused}

\item{verbose}{add verbose information}
}
\description{
\code{fortify_cellpop} prepares plottable data.frame.

\code{plot_driver} draws tbl_tree.
}
