% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{fortify_cellpop}
\alias{fortify_cellpop}
\alias{plot_tree}
\alias{geom_nodename}
\alias{geom_driver}
\title{Functions for visualization}
\usage{
fortify_cellpop(model, data)

plot_tree(data, ...)

geom_nodename()

geom_driver()
}
\arguments{
\item{model}{output of \code{scale_branches()}}

\item{data}{tbl_tree}

\item{...}{passed to \code{ggtree::geom_tree()} aes mapping}
}
\description{
Functions for visualization
}
\details{
\code{fortify_cellpop} prepares plottable data.frame.

\code{plot_tree} draws tbl_tree.

\code{geom_nodename} annotates tree nodes.

\code{geom_driver} annotates driver mutations.
}
