% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_branches}
\alias{scale_branches}
\alias{add_extra_columns}
\alias{filter_scale_tips}
\alias{nest_tippairs}
\alias{unnest_children}
\title{Scale branches}
\usage{
scale_branches(x)

add_extra_columns(x)

filter_scale_tips(x)

nest_tippairs(x)

unnest_children(x)
}
\arguments{
\item{x}{data.frame}
}
\description{
\code{scale_branches} is a shortcut

\code{add_extra_columns} prepares necessary columns for scaling
}
