% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_branches}
\alias{scale_branches}
\alias{scale_branches_record}
\alias{filter_scale_tips}
\alias{nest_tippairs}
\alias{unnest_children}
\title{Scale branches}
\usage{
scale_branches(x, detector = detect_driver_pois)

scale_branches_record(x)

filter_scale_tips(x, detector)

nest_tippairs(x, detector)

unnest_children(x)
}
\arguments{
\item{x}{data.frame}

\item{detector}{function}
}
\description{
Scale branches
}
\details{
\code{scale_branches} is a shortcut.

\code{scale_branches_record} records scaling process for explanation.
}
