% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{infer_rooted_tree}
\alias{infer_rooted_tree}
\title{Infer rooted tree of sampled genotypes}
\usage{
infer_rooted_tree(genotypes)
}
\arguments{
\item{genotypes}{string vector}
}
\value{
tibble
}
\description{
Infer rooted tree of sampled genotypes
}
