% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{as_phylo}
\alias{as_phylo}
\alias{as_multiphylo}
\alias{group_clade}
\alias{remove_outgroup}
\title{Take normal data.frame as tbl_tree for simplicity}
\usage{
as_phylo(x)

as_multiphylo(tbls)

group_clade(x, node)

remove_outgroup(x)
}
\arguments{
\item{x}{tbl_tree}

\item{tbls}{list of tbl_tree}

\item{node}{ancestral node}
}
\description{
Take normal data.frame as tbl_tree for simplicity

Add class name

Remove dummy outgroup
}
