% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{as_phylo}
\alias{as_phylo}
\alias{as_multiphylo}
\alias{group_clade}
\alias{upstream_corner}
\title{Functions to process tidytree objects}
\usage{
as_phylo(x)

as_multiphylo(tbls)

group_clade(x, node)

upstream_corner(x, node)
}
\arguments{
\item{x}{tbl_tree}

\item{tbls}{list of tbl_tree}

\item{node}{ancestral node}
}
\description{
\code{as_phylo} is a shortcut to convert data.frame to ape::phylo.

\code{as_multiphylo} adds class name.

\code{group_clade} is a shortcut to apply \code{tidytree::groupClade} to data.frame.

\code{upstream_corner} finds the joint between a node and its parent
}
