% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{infer_rooted_tree}
\alias{infer_rooted_tree}
\alias{as.phylo.data.frame}
\alias{as_multiphylo}
\alias{remove_outgroup}
\alias{rescale_children}
\title{Infer rooted tree of sampled genotypes}
\usage{
infer_rooted_tree(genotypes)

\method{as.phylo}{data.frame}(x)

as_multiphylo(tbls)

remove_outgroup(x)

rescale_children(x, scale)
}
\arguments{
\item{genotypes}{string vector}

\item{x}{tbl}

\item{tbls}{list of tbl}

\item{scale}{numeric}
}
\value{
tibble

phylo

multiPhylo

tibble

tibble
}
\description{
Infer rooted tree of sampled genotypes

Take normal data.frame as tbl_tree for simplicity

Add class name

Remove dummy outgroup

Rescale chilren branches recursively
}
